+++ CHANGELOG.md
@@ -7,18 +7,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Pagination on list view
+
 ## [0.2.0] - 2022-03-28
 
 ### Added
+
 - Assembly Option list view
 
 ## [0.1.0] - 2022-03-14
 
 ### Added
+
 - Assembly Option graphQL validation added to the Front-End code
 - Base Error and Success Messages defined
 
 ### Added
+
 - Base Assembly Options page and form created
 - Base Key and SKU sections added to the Assembly Form
 - All Intl messages added